Struct google_apigateway1::api::ProjectMethods

pub struct ProjectMethods<'a, C> where
    C: 'a,  { /* fields omitted */ }

A builder providing access to all methods supported on project resources. It is not used directly, but through the Apigateway hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_apigateway1 as apigateway1;
 
use std::default::Default;
use oauth2;
use apigateway1::Apigateway;
 
let secret: oauth2::ApplicationSecret = Default::default();
let auth = yup_oauth2::InstalledFlowAuthenticator::builder(
        secret,
        yup_oauth2::InstalledFlowReturnMethod::HTTPRedirect,
    ).build().await.unwrap();
let mut hub = Apigateway::new(hyper::Client::builder().build(hyper_rustls::HttpsConnector::with_native_roots()), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `locations_apis_configs_create(...)`, `locations_apis_configs_delete(...)`, `locations_apis_configs_get(...)`, `locations_apis_configs_get_iam_policy(...)`, `locations_apis_configs_list(...)`, `locations_apis_configs_patch(...)`, `locations_apis_configs_set_iam_policy(...)`, `locations_apis_configs_test_iam_permissions(...)`, `locations_apis_create(...)`, `locations_apis_delete(...)`, `locations_apis_get(...)`, `locations_apis_get_iam_policy(...)`, `locations_apis_list(...)`, `locations_apis_patch(...)`, `locations_apis_set_iam_policy(...)`, `locations_apis_test_iam_permissions(...)`, `locations_gateways_create(...)`, `locations_gateways_delete(...)`, `locations_gateways_get(...)`, `locations_gateways_get_iam_policy(...)`, `locations_gateways_list(...)`, `locations_gateways_patch(...)`, `locations_gateways_set_iam_policy(...)`, `locations_gateways_test_iam_permissions(...)`, `locations_get(...)`, `locations_list(...)`, `locations_operations_cancel(...)`, `locations_operations_delete(...)`, `locations_operations_get(...)` and `locations_operations_list(...)`
// to build up your call.
let rb = hub.projects();

Implementations

impl<'a, C> ProjectMethods<'a, C>

pub fn locations_apis_configs_create(
&self,
request: ApigatewayApiConfig,
parent: &str
) -> ProjectLocationApiConfigCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates a new ApiConfig in a given project and location.

Arguments

request - No description provided.

parent - Required. Parent resource of the API Config, of the form: projects/*/locations/global/apis/*

pub fn locations_apis_configs_delete(
&self,
name: &str
) -> ProjectLocationApiConfigDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes a single ApiConfig.

Arguments

name - Required. Resource name of the form: projects/*/locations/global/apis/*/configs/*

pub fn locations_apis_configs_get(
&self,
name: &str
) -> ProjectLocationApiConfigGetCall<'a, C>

Create a builder to help you perform the following task:

Gets details of a single ApiConfig.

Arguments

name - Required. Resource name of the form: projects/*/locations/global/apis/*/configs/*

pub fn locations_apis_configs_get_iam_policy(
&self,
resource: &str
) -> ProjectLocationApiConfigGetIamPolicyCall<'a, C>

Create a builder to help you perform the following task:

Gets the access control policy for a resource. Returns an empty policy if the resource exists and does not have a policy set.

Arguments

resource - REQUIRED: The resource for which the policy is being requested. See the operation documentation for the appropriate value for this field.

pub fn locations_apis_configs_list(
&self,
parent: &str
) -> ProjectLocationApiConfigListCall<'a, C>

Create a builder to help you perform the following task:

Lists ApiConfigs in a given project and location.

Arguments

parent - Required. Parent resource of the API Config, of the form: projects/*/locations/global/apis/*

pub fn locations_apis_configs_patch(
&self,
request: ApigatewayApiConfig,
name: &str
) -> ProjectLocationApiConfigPatchCall<'a, C>

Create a builder to help you perform the following task:

Updates the parameters of a single ApiConfig.

Arguments

request - No description provided.

name - Output only. Resource name of the API Config. Format: projects/{project}/locations/global/apis/{api}/configs/{api_config}

pub fn locations_apis_configs_set_iam_policy(
&self,
request: ApigatewaySetIamPolicyRequest,
resource: &str
) -> ProjectLocationApiConfigSetIamPolicyCall<'a, C>

Create a builder to help you perform the following task:

Sets the access control policy on the specified resource. Replaces any existing policy. Can return NOT_FOUND, INVALID_ARGUMENT, and PERMISSION_DENIED errors.

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy is being specified. See the operation documentation for the appropriate value for this field.

pub fn locations_apis_configs_test_iam_permissions(
&self,
request: ApigatewayTestIamPermissionsRequest,
resource: &str
) -> ProjectLocationApiConfigTestIamPermissionCall<'a, C>

Create a builder to help you perform the following task:

Returns permissions that a caller has on the specified resource. If the resource does not exist, this will return an empty set of permissions, not a NOT_FOUND error. Note: This operation is designed to be used for building permission-aware UIs and command-line tools, not for authorization checking. This operation may "fail open" without warning.

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy detail is being requested. See the operation documentation for the appropriate value for this field.

pub fn locations_apis_create(
&self,
request: ApigatewayApi,
parent: &str
) -> ProjectLocationApiCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates a new Api in a given project and location.

Arguments

request - No description provided.

parent - Required. Parent resource of the API, of the form: projects/*/locations/global

pub fn locations_apis_delete(
&self,
name: &str
) -> ProjectLocationApiDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes a single Api.

Arguments

name - Required. Resource name of the form: projects/*/locations/global/apis/*

pub fn locations_apis_get(&self, name: &str) -> ProjectLocationApiGetCall<'a, C>

Create a builder to help you perform the following task:

Gets details of a single Api.

Arguments

name - Required. Resource name of the form: projects/*/locations/global/apis/*

pub fn locations_apis_get_iam_policy(
&self,
resource: &str
) -> ProjectLocationApiGetIamPolicyCall<'a, C>

Create a builder to help you perform the following task:

Gets the access control policy for a resource. Returns an empty policy if the resource exists and does not have a policy set.

Arguments

resource - REQUIRED: The resource for which the policy is being requested. See the operation documentation for the appropriate value for this field.

pub fn locations_apis_list(
&self,
parent: &str
) -> ProjectLocationApiListCall<'a, C>

Create a builder to help you perform the following task:

Lists Apis in a given project and location.

Arguments

parent - Required. Parent resource of the API, of the form: projects/*/locations/global

pub fn locations_apis_patch(
&self,
request: ApigatewayApi,
name: &str
) -> ProjectLocationApiPatchCall<'a, C>

Create a builder to help you perform the following task:

Updates the parameters of a single Api.

Arguments

request - No description provided.

name - Output only. Resource name of the API. Format: projects/{project}/locations/global/apis/{api}

pub fn locations_apis_set_iam_policy(
&self,
request: ApigatewaySetIamPolicyRequest,
resource: &str
) -> ProjectLocationApiSetIamPolicyCall<'a, C>

Create a builder to help you perform the following task:

Sets the access control policy on the specified resource. Replaces any existing policy. Can return NOT_FOUND, INVALID_ARGUMENT, and PERMISSION_DENIED errors.

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy is being specified. See the operation documentation for the appropriate value for this field.

pub fn locations_apis_test_iam_permissions(
&self,
request: ApigatewayTestIamPermissionsRequest,
resource: &str
) -> ProjectLocationApiTestIamPermissionCall<'a, C>

Create a builder to help you perform the following task:

Returns permissions that a caller has on the specified resource. If the resource does not exist, this will return an empty set of permissions, not a NOT_FOUND error. Note: This operation is designed to be used for building permission-aware UIs and command-line tools, not for authorization checking. This operation may "fail open" without warning.

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy detail is being requested. See the operation documentation for the appropriate value for this field.

pub fn locations_gateways_create(
&self,
request: ApigatewayGateway,
parent: &str
) -> ProjectLocationGatewayCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates a new Gateway in a given project and location.

Arguments

request - No description provided.

parent - Required. Parent resource of the Gateway, of the form: projects/*/locations/*

pub fn locations_gateways_delete(
&self,
name: &str
) -> ProjectLocationGatewayDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes a single Gateway.

Arguments

name - Required. Resource name of the form: projects/*/locations/*/gateways/*

pub fn locations_gateways_get(
&self,
name: &str
) -> ProjectLocationGatewayGetCall<'a, C>

Create a builder to help you perform the following task:

Gets details of a single Gateway.

Arguments

name - Required. Resource name of the form: projects/*/locations/*/gateways/*

pub fn locations_gateways_get_iam_policy(
&self,
resource: &str
) -> ProjectLocationGatewayGetIamPolicyCall<'a, C>

Create a builder to help you perform the following task:

Gets the access control policy for a resource. Returns an empty policy if the resource exists and does not have a policy set.

Arguments

resource - REQUIRED: The resource for which the policy is being requested. See the operation documentation for the appropriate value for this field.

pub fn locations_gateways_list(
&self,
parent: &str
) -> ProjectLocationGatewayListCall<'a, C>

Create a builder to help you perform the following task:

Lists Gateways in a given project and location.

Arguments

parent - Required. Parent resource of the Gateway, of the form: projects/*/locations/*

pub fn locations_gateways_patch(
&self,
request: ApigatewayGateway,
name: &str
) -> ProjectLocationGatewayPatchCall<'a, C>

Create a builder to help you perform the following task:

Updates the parameters of a single Gateway.

Arguments

request - No description provided.

name - Output only. Resource name of the Gateway. Format: projects/{project}/locations/{location}/gateways/{gateway}

pub fn locations_gateways_set_iam_policy(
&self,
request: ApigatewaySetIamPolicyRequest,
resource: &str
) -> ProjectLocationGatewaySetIamPolicyCall<'a, C>

Create a builder to help you perform the following task:

Sets the access control policy on the specified resource. Replaces any existing policy. Can return NOT_FOUND, INVALID_ARGUMENT, and PERMISSION_DENIED errors.

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy is being specified. See the operation documentation for the appropriate value for this field.

pub fn locations_gateways_test_iam_permissions(
&self,
request: ApigatewayTestIamPermissionsRequest,
resource: &str
) -> ProjectLocationGatewayTestIamPermissionCall<'a, C>

Create a builder to help you perform the following task:

Returns permissions that a caller has on the specified resource. If the resource does not exist, this will return an empty set of permissions, not a NOT_FOUND error. Note: This operation is designed to be used for building permission-aware UIs and command-line tools, not for authorization checking. This operation may "fail open" without warning.

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy detail is being requested. See the operation documentation for the appropriate value for this field.

pub fn locations_operations_cancel(
&self,
request: ApigatewayCancelOperationRequest,
name: &str
) -> ProjectLocationOperationCancelCall<'a, C>

Create a builder to help you perform the following task:

Starts asynchronous cancellation on a long-running operation. The server makes a best effort to cancel the operation, but success is not guaranteed. If the server doesn't support this method, it returns google.rpc.Code.UNIMPLEMENTED. Clients can use Operations.GetOperation or other methods to check whether the cancellation succeeded or whether the operation completed despite cancellation. On successful cancellation, the operation is not deleted; instead, it becomes an operation with an Operation.error value with a google.rpc.Status.code of 1, corresponding to Code.CANCELLED.

Arguments

request - No description provided.

name - The name of the operation resource to be cancelled.

pub fn locations_operations_delete(
&self,
name: &str
) -> ProjectLocationOperationDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes a long-running operation. This method indicates that the client is no longer interested in the operation result. It does not cancel the operation. If the server doesn't support this method, it returns google.rpc.Code.UNIMPLEMENTED.

Arguments

name - The name of the operation resource to be deleted.

pub fn locations_operations_get(
&self,
name: &str
) -> ProjectLocationOperationGetCall<'a, C>

Create a builder to help you perform the following task:

Gets the latest state of a long-running operation. Clients can use this method to poll the operation result at intervals as recommended by the API service.

Arguments

name - The name of the operation resource.

pub fn locations_operations_list(
&self,
name: &str
) -> ProjectLocationOperationListCall<'a, C>

Create a builder to help you perform the following task:

Lists operations that match the specified filter in the request. If the server doesn't support this method, it returns UNIMPLEMENTED. NOTE: the name binding allows API services to override the binding to use different resource name schemes, such as users/*/operations. To override the binding, API services can add a binding such as "/v1/{name=users/*}/operations" to their service configuration. For backwards compatibility, the default name includes the operations collection id, however overriding users must ensure the name binding is the parent resource, without the operations collection id.

Arguments

name - The name of the operation's parent resource.

pub fn locations_get(&self, name: &str) -> ProjectLocationGetCall<'a, C>

Create a builder to help you perform the following task:

Gets information about a location.

Arguments

name - Resource name for the location.

pub fn locations_list(&self, name: &str) -> ProjectLocationListCall<'a, C>

Create a builder to help you perform the following task:

Lists information about the supported locations for this service.

Arguments

name - The resource that owns the locations collection, if applicable.